requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.5.3
coveralls==1.6.0
docopt==0.6.2
flasgger==0.9.2
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==6.0.0
# pkg-resources==0.0.0
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.7.1
pyrsistent==0.14.11
pytest==4.3.1
pytest-cov==2.6.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
